BUG24699422 Adding README and LICENSE files to Windows installation
Currently the MySQLRouter Windows installer created via cpack command copies
README.txt and Licence.txt files to the installation doc subdirectory.
Those files are missing when the installer is created with the msbuild.

Added the copying of the documentation files to the mysql_router.wxs.in file
which makes them appear in the doc directory when the installer is built using
msbuild.