Use --name for cache key
Append the name of a router instance to the metadata_cache config
entry, so that it can be used to differentiate between multiple cache
configurations. Combined with location param in routing URI, this allows
support for multiple cluster configurations in the same router.
However bootstrap continues supporting a single instance.

Support for --conf-use-sockets

Improved unit-tests.