fixed SIGABRT on solaris
terminate called after throwing an instance of 'std::out_of_range'
  what():  basic_string::at: __n (which is 1) >= this->size() (which is 1)

triggered by thread_id.at(1) assuming that the string is at least
2-bytes long.

But if the thread-id is only one digit, the exception will be triggered
and then lead to the abort().

The 0x prefix that this code-snippet tries to normalize is added
depending on the locale. Using std::noshowbase forces the same
behaviour on all platforms.