Bug#24807941 Bug#24808752
BOOTSTRAP COMMAND DOESNT ADD PROTOCOL OPTION IN MYSQLROUTER CONFIG
FILE

Add protocol=classic entry in classic protocol routing rules,
instead of relying on default.

INVALID INPUTS ARE NOT HANDLED FOR --NAME OPTION WHILE BOOTSTRAPPING THE
ROUTER

Changed use of the --name option. The router name is now an optional
value which is used to identify the router instance and can have any
characters other than newlines.

The cluster name is now used as the key for metadata_cache sections, to
allow future support for multiple clusters in the same configuration.
The cluster name is also used, in addition to the replicaset name and
the routing rule type, as the key for routing sections.

Attempting to overwrite a config file with a different cluster will
throw an error. The --force option is required to overwrite.