Fix for Bug #24955339 ROUTING TO DEFAULT PORT FOR X PROTOCOL NOT WORKING
There was still one place in the code where server port was hard-coded to 3306 -
classic protocol port.
I've fixed that and added UTs in the functionality around the code with the bug.
I've also did a small rework to pass the protocol type to different objects not
by string but using enum type.