Fix for Bug #25035285 ERROR WHILE MAKING SSL CONNECTION USING MYSQLXTEST CLIENT.
The logic in the x-protocol routing verifies the first messge that is send by the client.
It invalidates the connection if it is not one of the expected messages.
These expected messages were AuthOk and CapabilitiesGet.
When making SSL connection the first message that mysqlxtest sends is CapabilitiesSet
which is also allowed by the protocol.

I've added CapabilitiesSet and ConnectionClose to the expected messages set.