Fix for Bug #25048153 ROUTER DOESN'T COMPILE ON SOME PLATFORMS BECAUSE OF READDIR_R DEPRECATED WARNING
On some platforms (like Ubuntu 16.10 or Fedora 25) the compilation of the
router fails with:

error: 'int readdir_r(DIR, dirent, dirent**)' is deprecated
[-Werror=deprecated-declarations]

In the latest libc readdir_r has being depracated, which leads to this warning being issued and
breaking the compilation.

I've kept the usage of readdir_r as we still want compatibity with older library versions
where readdir was not thread safe.
I've added suppression for that warning.