Fix for Bug #25033421 ROUTER TERMINATES WHEN THE SOCKET IT WRITES TO GETS CLOSED.
There is the issue that I oberve on my Ubuntu16 almost each time when I run:

mysqlsh --uri=root:@127.0.0.1:7001  -e 'select 1;'

In this scenario the shell is first trying x-protocol connection on classic protocol socket.
The router sees that and does not approve.
It tries to send an error but by that time the shell already cloeses that socket.
That leads to the signal being sent to the router from the os that terminates the router.
Reproduction rate is nearly 100% at my system.

Registering ignore-handler for the SIGPIPE fixes this.