Fix for Bug #25033533 ROUTER SENDS INVALID DATA TO CLIENT IN CASE OF ERROR DURING HANDSHAKE IN CLASSIC
During the handshake phase router inspects the packets exchanged between the
client and MySQL Server.
When the Server responds with an error the Router creates the ErrorPacket
object that gets send to the client.
The problem is that this object gets created with the constructor taking the
whole buffer 16kB (that works as an application layer network buffer between the client and
server). This causes to attempt writing to the socket the whole buffer (16kB) even
though only beginning of it contains serialized error.

Fixed by creating the packed from copy of the part of the buffer that contains the read data.