Fix for bug 25069252 / MYR-269 ROUTING NOT WORKING WHEN ROUTER STARTED AS A SERVICE
When the router runs as windows service std::getline (and other console IOs) returns "".

Solution is to log an error to windows event log and abort when console interaction is required (boostrap and running with metadata_cache without master key file). Some stages where the error condition happen are before plugins are initialized, thus the logger plugin is not available yet.
Also a popup a message as error wont be good, since service can be started with no session active in the machine.
The current implementation for writing to windows log is provisional: once Mats's refactoring of logging is merged and the real plugin for windows event log is done, another approach will be considered.

rb14601.alfredo