Fix for Bug #24962552 CORE DUMP WHEN ROUTER STARTED TCP AND SOCKET CONFIGS USING --A OPTION
The router (sometimes) crashes or generally behaves unpredictably when
there is more than one section for metadata_cache in the configuration.
That leads to multiple instances of metadata_cache classes being
created which is not supported (because of the globals etc.)

I've added a mechanism that checks for such invalid configuration, warns
the user and closes the router.