Bug #24673606 - pause connections until failover
FIRST CLIENT CONN FAILS FOR R/W MODE,WITH DEFAULT TTL
ONCE THE PRIMARY IS DOWN

When a primary server dies, it may take over 6 seconds for a new primary
to be elected. Furthermore, because the metadata is refreshed once every
TTL seconds, the router will only notice the failover of the primary
after that amount of time.

If a connection arrives before that time, a refresh will be triggered
and the new primary will be reflected in the metadata cache, but the
connection that triggered the refresh will fail. That 1st conenction
that triggered the refresh will fail even if the failover had already
happened, because the metadata cache still has the old view of the
cluster.

This patch fixes the issue by blockin connections that notice that the
primary has died, making them wait for up to X seconds (10 by default)
for the refresh to happen. If during that time, a new primary is seen in
the metadata, the connection will be unblocked and succeed. If no
primary is elected even after that timeout, the connection will fail as
usual.

Additionally:
- added a ManagedReplicaSet object, which encapsulates metadata from a
  replicaset (like whether it's single-primary mode and in the future,
things like the group_id and its view_id), in addition to the list of
instances
- GRANT privileges to PFS.replication_group_member_status table during
  bootstrap, to allow querying the group_id and view_id