Fixed Solaris bug and test failures
struct dirent is not supposed to be allocated statically, it must be
allocated taking the max length of a directory entry into account. In
Linux, it happens that the dirent.d_name field has enough space
built-in, but in Solaris d_name has length 1 which leads to a buffer
overrun.

Also in Solaris, /etc/services is pretty empty and is missing entries
for http, https and mysql. Changed the tested services to more
old-school things like ssh and ftp. Left mysql out, like it is in
Windows.

Finally, open(2) on a directory gives a different error than in Linux.