Fix for gtest tests crashing when compiled with gcc-6.
There is a known bug:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=833450
causing the mock call verification failure on some scenarios when gcc-6 is used.
We hit it in our test_metadata_cache testcase (it fails on pb2 on the Linux distros using fresh gcc, like ubuntu 16.10)

I've applied the patch that fixes the bug in our bundled gtest version.
Hovewer the user still can use it's own gtest (WITH_GMOCK cmake option).
We prevent that in case gcc-6 or greater is used.