The existing test_bug22020711.cc was doing some integration level stuff like launching the whole router as a separate process, talking to it and then killing it providing its pid. The killing part sometimes failes on pb2.
All that was done to verify if the proper error is being sent from the router
to client if no destination servever is available. I've reworked the testcase
to be more unti test-like so there is no spawnig the process, no killing, etc.
I also moved it from issues/test_bug22020711.cc to test_classic_protocol.cc

rb#14790