Fix for MYR-357 - Unit test KeyringManager.init_with_key_file failing for windows sporadically
Fixes for sporadic fails of unit tests of KeyringManager suite.
- The change in file contents loading in test_keyring_manager.cc was required to reduce the failure rate from 20% to 1% (all errors on the tests  check_kf.check_unchanged() & check_kr.check_unchanged()). Other than that, there was nothing wrong with the previous approach.
- The 2nd fix, to retry saving keyring file, was to fix another error that appeared after fixing the previous one, was causing failure in rate of 1%.
  (As explained in the comments, this was likely caused by the indexer service in windows, as when such option was disabled for the build folder where the tests were ran (the checkbox  "Allow files in this folder to have contents indexed in addition to file properties"), the failure disappeared). The fix was tested with the indexing reenabled.

Tests dpne:
With the fix, it seems to work 100% of times (ran a few times the KeyringManager suite, 100 times each, and always passed all, also ran the whole suite of router tests and they all passed).

Originally suite was failing between 15%-20% of the runs, this could be verified with the following

REM Iterates 100 times the KeyringManager test
FOR /L %%A IN (1,1,100) DO (
  ctest -C Debug -V -R KeyringManager >> results.txt
)

UPDATE: Fix currently pushed to build/dev2 for testing in pb2.

rb14808.andrzej