Fix for BUG 25045182 - DEFAULT KEYRING PATH CAN LEAD TO KEYRING LOSS
Currently the default location for the keyring file is runtime folder.
On some installation layouts it is /var/run/ which is temporary and
does not survive machine reboot.

Added a notion of data folder which default to /var/lib on those
system. The idea is to use it for any persisent data that the router
needs to store.