WL#9859 Option to run as a cmdline specified user
To enable a secure deployment of the router, it is necessary to allow it to run
as  an user with restricted system privileges. By running the router under such
a user, a compromised router would have limited impact on the host.
The router should have an option to change its user id during the startup
similar to --user option in mysqld.