Fix for Bug #25468232 HEAP-BUFFER-OVERFLOW IN BOOTSTRAP IF SCHEMA_VERSION HAS WRONG STRUCTURE
The code inside router/src/cluster_metadata.cc does various sql queries to check
for validity of the cluster.
The number of the values returned from those checks is not always verified
before accessing those values. In other places assert get used which will also not
prevent the crash in the runtime.

Added the exceptions (in place of existing asserts when those were used) whenever
number of the fields returned from the query differs from the expected value.