Fix for Bug #25460958 ROUTER DOESN'T THROW ANY ERROR FOR INVALID METADATA_CACHE USER
This bug is closely related to the plugin lifecycle work that is on hold
and reveals the problem we have with the current design:
If the plugin throws some exception and does not handle it inside of the
plugin it gets catched in the Loader.start_all() method. The Loader then tries
to call stop_all on all the loaded plugins and waits for them to stop.
The problem is that none of the plugins implements stop() method in a
correct way (it should cause all the plugin's threads to be stopped and leave to
the caller). The effect is that after such exception is thrown the throwing
plugin does not operate, the loader hangs in the start_all() method and
there is not even a trace of any error in the log file.

I did not fix the generic problem as this is supposed to be solved with
the plugin lifecycle WL. I only added a handling to this particular
exception, but generally the problem is much more complex (see above).