Fix for Bug #25450178 BOOTSTRAPPING ROUTER SHOULD BE ATOMIC
This fix probably does not solve all non-atomic scenarios for
bootstrap possible
(we will have separate WL for that if needed) but fixes the issue
described directly in the bug report. When existing file (not directory)
was given as and option to bootstrap -d, and it was mixed with --user
option, the bootstrap was changing the ownership of this file before
giving an error that it can't do bootrap using this file as an output
directory.

Also fixed minor issue in the harness that surfaced with this bug:
wrong usage of strerror_r() that caused never printing the message
for the errno os the systems where strerror_r returns char* instead
of int. Used existing wrapper to fix it.