Fix for Bug #25493968 STARTING TWO ROUTERS WITH SAME CONFIG FILE LEADS TO SEG-FAULT
The problem is caused by the fact that we get to
the end of main() function and call mysql_library_end() while there
is still unjoined thread in the metadata-cache plugin
that is being joined only in the destructor of the global
variable g_metadata_cache. This thread uses mysql-lib
that we deinitialize which leads to the crash.

Until we introduce WL9558 Plugin life-cycle that support graceful
shutdown and restart. the simple solution is to skip mysql-lib
deinitialization.