Fix for Bug #25525481 ROUTER SERVICE ON WINDOWS DOESN'T WORK
The code in Path::real_path() after resolving
the absolute file path calls FindFirstFile on
that path to make sure the file exists.
The problem is that this call can yield ERROR_ACCESS_DENIED but
that does not necessarily mean that the path does not exist.
We still can have the access to the
file itself but we can't call the Find on the directory that
contains the file.
This is true for example when the config file is placed in
the User's directory and it is accesseed by the router that
is run as a Windows service.)

Fixed by treating ERROR_ACCESS_DENIED as if the path existed.
Also disabled for Windows the code that was creating empty log file
during the bootstrap. This is only needed for --user option which
is not supported on Windows and was causing the log file being
sometimes not accessible to the Windows service if it was created
this way.