Fix for Bug 25549490 UNIT TEST 'BAD SUKI' FAILS ON FREEBSD BUILDS
The failing test defines plugin-specific exception that gets thrown
outside of the plugin (so library). To make the exception type known to the
test binary it is being declared and defined in the header file. That leads
to violation of the ODR. On the FreeBSD that exception called 'bad_suki'
misses the catch(bad_suki) clause as this is seen as another type in the
test executable outside the so library.

The test has a couple of flaws: the bigest of them is expecting the calling code
which is the framework to know about the type of the exception that is specific to
some plugin. That inverts the dependency as the framork should not depend on the
plugin.

I've fixed it by catching the exception by the parent type which is std::runtime_error
and is know for both plugin and loading code.