Fix for Bug #25522880 MYSQLROUTER FAILS TO RUN ON SOLARIS - REFERENCED SYMBOL NOT FOUND
The reason for this problem is that on Solaris the Router is build with
newer C++ standard library version than the one that is in the Solaris
system directory.

MySQLServer bundles libstdc++.so to address this problem. I've ported
the cmake magic that does that to the Router. The code is not changed -
moved directly from server's configure.cmake.