Fix for Bug #25682736 LINUX BOOTSTRAP GENERATED CONFIG FILE SHOULD BE WORLD READABLE
If the user does the bootstrap with superuser (uid==0) but did not provide
--user option [s]he will end up up with the files (config, log, etc.) owned
by the root user and not accessible by others, which is likely not what was expected.

Changed by forcing the user to use --user option when bootstraping as superuser.
The user still can use --user=root to force using superuser.