Fix for Bug #25688333 LINUX SERVICE FILES AND MYSQLROUTER BINARY USE DIFFERENT DEFAULT CONFIG FILES
Removed all the references to/instances of the *.ini files in the router.
Replaced with *.conf.

Added functionality that fallbacks to ini file if default file with conf
extension is to be used.
This is for backward compatibility with previous reouter installations
as ini extension was used by the previous versions.