Harness tests now use DLL/SO version of Harness library
We build two versions of Harness library: static and dynamic.
Before this patch, harness unit tests would use the static version, now
they use the dynamic version.  This change was necessary, because of the
following scenario:

  Step 0. unit test uses static Harness library
  Step 1. unit test loads a plugin via Harness' Loader
  Step 2. that plugin needs to call a Harness function, but since it's built
          to use the dynamic version of Harness library, it loads that
  The effect: we have 2 versions of Harness library loaded, one static
  and one dynamic.

This hasn't been a problem up until now, but surfaced with our change to
new logger, which is now inside the Harness library.  Thus the test and
the plugin will call log_*() from two different libraries.  This is a
problem, because the new logger hold state.  Particularly, it needs to
be initialized.  We do this initialization in unit test prior to running
the plugin, but this results in initializing the logger in static
library, while the plugin calls the uninitialized logger in the dynamic
library.