fixed the is-running-task running check
start-stop-daemon --exec --chuid fails with EPERM on /proc/<pid>/exe
which leads to the 'start' not seeing the process already running even
though a --pidfile is provided, and --stop fails to stop as it thinks
the process isn't running.

Using --startas bypasses the /proc/<pid>/exe check and makes it all
work.

Also removed the apparmor dependency as we don't have a hard dependency
on it.

fixes #25826607, #25825821 partially