improved error-checking and handling during plugin loading
Before this patch, we relied on checking Plugin* obtained from
dlsym()/GetProcAddress(): if they returned NULL, the operation was
assumed to have failed. While this worked for our case (because our
Plugin* would never be NULL), in general it should be done differently:
dlerror()/GetLastError() should be queried. This patch changes that.

Also, the code now outputs the actual error message obtained from the
system on error.