Fix for Bug #25956532 MYSQLROUTER DOES NOT START WITH DEFAULT MYSQLROUTER --BOOTSTRAP
The bootstrap procedure for the system-wide bootstrap creates the keyring
directory in case it does not exist (it should be there after the installation
but still it can be removed later on). This directory did not get the proper
ownership set in case the bootstrap was ran with --user option.
As a result the router would not start after the bootstrap due to the access
rights issue.

Fixed by setting the newly created directory ownership to the --user.