Fix for Bug #26044181 Router reconnect time too long
The problem is with the inproper usage of the mutex that led
to 10s sleep for each connection while keeping the mutex.
That led to all other connection threads waiting for the mutex
and sleeping 10s one after another meaning N * 10s reconnect
for N = [num of connections].

Fixed by narrowing the mutex lock to protect only the required
shared resource (current_pos_ iterator).