WL#10822 Revised plugin lifecycle and added clean shutdown capability: add Windows support
This patch is a continuation of WL#9558, but this time it's for Windows:

  a30eef9 (2017-06-01 08:33:40 +0200) Pawel Mroszczyk:
  WL#9558 MYR-97 Revised plugin lifecycle and added clean shutdown capability (unix only)

Please see the comments in the above commit to understand the purpose.

Officially, previous WL supported only Unix platforms, but in reality it
also supported Windows in cases when the shutdown was initiated
internally (triggered by an error or natural shutdown).  So this patch
only adds the other case: when shutdown is triggered by an external
event (such as user pressing Ctrl+C or telling Router Window Service to
stop).

Notable changes in this patch:

  - added 2 ways to shut down on Windows:
    - pressing Ctrl+C when running from DOS shell
    - clicking STOP on "MySQL Router" in Windows Services window
      (or its equivalent command in DOS shell: net stop "MySQL Router")

  - fixed and enabled related unit tests in Windows

  - fixed CMake problem: path to WiX Toolset was not always found by the
    CMake script

  - added serveral log_info() lines in Loader's plugin-loading step