BUG#26441587 fixed Router fails to start with config file having [logger] section
This patch is a (hopefully temporary) hack until a more proper solution
is implemented to provide backward compatibility of configuration files.
Before we introduced new logging infrastructure in WL#8923, our logger
was a separate plugin, and thus config files contained a [logger] section.
Now the logger is an integral part of the Harness and logger.so/dll no
longer exists, therefore it would seem that [logger] section should no
longer appear in the config file. Yet, we need to maintain backward
compatibility of the config file, therefore must allow [logger] section
to appear and carry the information it carried before.

This patch provisions this requirement.