fixed BUG#25853768 (start.sh contains bogus sudo reference); fixed stop.sh
This patch fixes start.sh generated during bootstrap, so that it does
not require sudo access when user running the script is the --user
specified in the bootstrap.

Also, stop.sh gerenated during bootstrap now uses the correct signal to
shut down Router, and will not erase PID-file if sending signal fails.