BUG#25111926 [keyring] Improved error messages on bad keyring files
The original Bug report was about exception being thrown on Unix and not
on Windows, but it turned out that it's also thrown on Windows, and
furthermore, this is correct behavior. It was however agreed, that the
error messages returned back to the user should be improved, and this
patch does that.