[routing plugin] BUG#BUG26800200: fix segfault when creating client thread fails
Before this patch, if creating a new client thread failed, std::thread
constructor would throw a exception which we would not handle. Now we
handle this situation gracefully and return an appropriate error to the
client via MySQL protocol. We also log this as an error, but only once
(to prevent generating heavy disk usage when this happens).