[server mock] BUG#27551925 Fix mock server not connecting to 5.7 clients
Before this patch, our mock server could only connect to clients using
libmysqlclient >= 8.0.4. This is because of the fast-auth step in the
handshake that we added to make it possible to connect the new 8.0.4(+)
clients. This patch restores ability to also connect to 5.7 clients.

Now the mock detects which client is connected to it and acts
accordingly. More precisely, it uses mysql_protocol library to decode
the handshake response message and depending on auth-plugin specified,
it performs fast-auth or or "classic" auth which the 5.7 clients expect.