Fix for 27721898 Bug #27721898 EXECUTING STATEMENTS FAILED WITH: 'UNKNOWN ERROR-CODE'
When the executing the SQL commands during the bootstrap resulted in an error
different than a few errors that we handle specifically we reported
'Unknown error' instead of the original error message that we got from the server.
Fixed by removing our versions of the messages for the server error codes, always using the message got from the server.