Fix for Bug #27721834 ERROR: THE PROVIDED SERVER CONTAINS AN UNSUPPORTED INNODB CLUSTER METADATA
Doing the queries during the bootstrap the common pattern we did was to just
bury the unexpected error (if happened) from the server and pretend to the caller that
something comepletely different has happened.
Like for example checking for the schema version, if the query to cluster returns Access denied
error we just catch the exception and return false, as if the schema version was
wrong. This is completely misleading to the caller (see the bug for details).

As a fix I removed all the catch-ing of unexpected errors during the queries
to let the original errors to bubble up to the caller and effectively to the user.