Fix BUG#28059228: Misleading log messages (and variable names + comments in code)
Before this patch, we had a bunch of stale code (variable names and
comments) and log messages that suggested that "emeregency mode"
(temporary increase of refresh frequency to 1/s) was triggred by PRIMARY
node going missing and disabled once new one was established. In
reality, this mechanism also applied to SECONDARY nodes, so this patch
fixes the code and log messages to reflect that.