Fix for Bug #28082473 [MDC] REFRESH: IF METADATA SQL QUERY FAILS, WHOLE REFRESH FAILS
This issue is not directly related to  WL#11954 but its severity is
higher now as we could be disconnecting
existing connections if that fails. Hence we should get it fixed for
that WL.

Now when refreshing the metadata if querying for the md on some md
server fails we try the others.